requirements.txt
langchain
langchain-community
langchain-openai
python-dotenv
openapi-spec-validator





